Did a few things. 1) Corrected a threading problem in the InputFileDataProvider which prevented us from having multiple of these reading in parallel. In the start method, we were actually waiting on the underlying thread to complete using the latch. This was probably originally put in place for the unit tests so we know when the reader is complete before checking results. Corrected this with a separate method to wait on completion.
2) Modified the logic in InputFileDataProvider and FileReader to close
the output stream in the oStrm/iStrm pipeline when we finish reading our
data from the file. This pipeline is used for passing data in the
NMEASentenceProvider sentence readers. By closing this output stream,
the inputstreams used by the sentence readers get properly closed. This
eliminated exceptions I was seeing in the sentence readers when they got
to the end of the input stream.

3) Modified DataProvider so it has an ID field, and then
NMEASentenceProvider to set this field. This allows this ID to be used
in thread names in the various data providers and sentence readers so
the correspond to each other. Makes it easier to understand the logs
when the FileReader0 feeds data in to the SentenceReader0, etc.